Refactor components to make smaller sized components
